feat(agent): implement subagent dispatch logic in a new module
- Added a new `dispatch.rs` file containing the `dispatch_subagent` function, which handles the logic for delegating tasks to subagents.
- The function includes error handling for uninitialized agent registries and unknown agent IDs, returning appropriate tool results.
- Integrated global event publishing for subagent lifecycle events, enhancing observability.
- Updated `mod.rs` to include the new dispatch module, improving code organization and maintainability.
- Added unit tests to ensure the correctness of the dispatch logic and error handling.
